docs(peagen): document generated task schemas
